agent_backend/pyproject.toml
[project]
name = "agent-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "claude-agent-sdk>=0.1.27",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "openai>=2.20.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.34.0",
]
